refactor(highlight-pattern): improve code
