+++ feature/matcher/src/commonMain/kotlin/com/neoutils/neoregex/feature/matcher/MatcherScreen.kt
@@ -37,19 +37,15 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.focus.focusProperties
 import androidx.compose.ui.focus.onFocusChanged
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.input.key.onPreviewKeyEvent
-import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import cafe.adriel.voyager.core.model.rememberScreenModel
 import cafe.adriel.voyager.core.screen.Screen
-import com.neoutils.highlight.compose.extension.spanStyle
 import com.neoutils.highlight.compose.remember.rememberTextFieldValue
-import com.neoutils.highlight.core.Highlight
 import com.neoutils.neoregex.core.common.util.Command
 import com.neoutils.neoregex.core.designsystem.textfield.NeoTextField
 import com.neoutils.neoregex.core.designsystem.theme.NeoTheme.dimensions
@@ -62,6 +58,7 @@ import com.neoutils.neoregex.core.sharedui.component.TextEditor
 import com.neoutils.neoregex.feature.matcher.action.MatcherAction
 import com.neoutils.neoregex.feature.matcher.extension.onLongHold
 import com.neoutils.neoregex.feature.matcher.extension.toTextState
+import com.neoutils.neoregex.feature.matcher.model.Syntax
 import com.neoutils.neoregex.feature.matcher.model.Target
 import com.neoutils.neoregex.feature.matcher.state.MatcherUiState
 import com.neoutils.neoregex.feature.matcher.state.error
@@ -295,117 +292,5 @@ class MatcherScreen : Screen {
                     horizontal = dimensions.small,
                 )
         )
-    }
-}
-
-interface Syntax {
-
-    val highlight: Highlight
-
-    class Regex(
-        private val modifierColor: Color,
-        private val escapeReservedColor: Color,
-        private val escapedCharColor: Color,
-        private val anchorsColor: Color,
-        private val charSetColor: Color,
-        private val groupColor: Color,
-        private val controlsColor: Color,
-        private val literalColor: Color
-    ) : Syntax {
-
-        private val charSetRegex = """(\\{2})|(\\\[)|(\[\^?)((?:\\{2}|\\\]|[^\]])*)(\]?)""".toRegex()
-        private val groupRegex = """(\\{2})|(\\\()|(\((?:\?[:=!])?)((?:\\{2}|\\\)|\\\[|\[.*\]|[^\)])*)(\)?)""".toRegex()
-        private val quantifierRegex = """(\\{2})|(\\\{)|(\{\w,?\w?\})""".toRegex()
-        private val escapeReservedRegex = """(\\{2})|(\\[{}()\[\]$^+*?])""".toRegex()
-        private val escapedCharRegex = """(\\{2})|(\\[DdWwSsHhVvR])""".toRegex()
-        private val anchorsRegex = """(\\{2})|(\\[$^])|(\\[AZzBbG])|([$^])""".toRegex()
-        private val controlsRegex = """(\\{2})|(\\[tnfrae])""".toRegex()
-        private val modifierRegex = """(\\{2})|(\\[+*?])|([+*?])""".toRegex()
-
-        private val charSetSpanStyle = SpanStyle(
-            color = charSetColor,
-            background = charSetColor.copy(
-                alpha = 0.2f
-            )
-        )
-
-        private val groupSpanStyle = SpanStyle(
-            color = groupColor,
-            background = groupColor.copy(
-                alpha = 0.2f
-            )
-        )
-
-        override val highlight = Highlight {
-
-            spanStyle {
-
-                groupRegex.match {
-
-                    // charset
-                    put(3, groupSpanStyle)
-                    put(4, groupSpanStyle.copy(color = Color.Unspecified))
-                    put(5, groupSpanStyle)
-                }
-
-                modifierRegex.match {
-
-                    // modifier
-                    put(3, SpanStyle(color = modifierColor))
-                }
-
-                quantifierRegex.match {
-
-                    // modifier
-                    put(3, SpanStyle(color = modifierColor))
-                }
-
-                charSetRegex.match {
-
-                    // charset
-                    put(3, charSetSpanStyle)
-                    put(4, charSetSpanStyle.copy(color = literalColor))
-                    put(5, charSetSpanStyle)
-                }
-
-                controlsRegex.match {
-
-                    // controls
-                    put(2, SpanStyle(color = controlsColor))
-                }
-
-                escapedCharRegex.match {
-
-                    // escaped char
-                    put(2, SpanStyle(color = escapedCharColor))
-                }
-
-                escapeReservedRegex.match {
-
-                    // full match
-                    put(0, SpanStyle(color = escapeReservedColor))
-                }
-
-                anchorsRegex.match {
-
-                    // anchors
-                    put(3, SpanStyle(color = anchorsColor))
-                    put(4, SpanStyle(color = anchorsColor))
-                }
-            }
-        }
-
-        companion object {
-            val Default = Regex(
-                modifierColor = Color(0xff0077ff),
-                escapeReservedColor = Color(0xffb700ff),
-                escapedCharColor = Color(0xfff5cd05),
-                anchorsColor = Color(0xffb06100),
-                charSetColor = Color(0xffe39b00),
-                groupColor = Color(0xff038d00),
-                controlsColor = Color(0xffff00c9),
-                literalColor = Color.White
-            )
-        }
     }
 }
